Overhauled the Threshold system to work within the limitations of the binary system instead of using the entire float value, resulting in much more consistent thresholds on the animation states.
Refactored and organized more code, cleaned up much of the surface code on the Script.

Changed the definition of binarySize to mean the number of bits rather than the Binary Resolution itself, which allows more flexibility in the code. Added writeDefaults toggle.

Added Parameter Memory Usage to clearly indicate how much space would be required to run the created parameter in the VRC Avatar Descriptor.

Added a small section below the Create Parameters which indicates what parameters to add to the Avatar Descriptor.