Tiny refactor of BinaryParamScript to house a new BinaryStateMachine class, and can now be instantiated as an object with set variables.
BinaryParamWindow handles the class appropriately. This was done to make adding features much easier and to prepare other tools to take advantage of the class.

Both now are contained under the VRCFaceTracking.EditorTools namespace. If you would like to use this script then you will need to be using this namespace, or be contained within it.